refactor(audit): move plan_persist into mvm_hostd::audit with cli shim
