+++ crates/mvm-hostd/src/audit/mod.rs
@@ -1,0 +1,5 @@
+//! Host-side audit binding: the chain-signed `AuditEmitter`, the host signing
+//! keypair, plan persistence, and the checkpoint bind helpers. Library API so
+//! both the CLI and fleet consumers emit identical chain entries.
+
+pub mod plan_persist;
